+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+<a name"0.4.2"></a>
+### 0.4.2 (2015-07-21)
+
+
+#### Bug Fixes
+
+* **VKI_CONFIG:** exposes sizeAdj configuration (issue #13) ([37e5230f](https://github.com/the-darc/angular-virtual-keyboard/commit/37e5230f))
+* **auto-close:** auto-close others VKIs when a new one is focused (issue #10) ([8bf63d74](https://github.com/the-darc/angular-virtual-keyboard/commit/8bf63d74))
+* **focus:** keep input focus when keyboard buttons (like shift, alt, caps, and others..) are ([f3b4f3c0](https://github.com/the-darc/angular-virtual-keyboard/commit/f3b4f3c0))
+
+
 <a name"0.4.1"></a>
 ### 0.4.1 (2015-05-11)